64 bit build fixed
git-svn-id: https://vistrpndsvn1.vi.vector.int/svn/CANoe_Components/OpenRTI/trunk@556777 fc9017fd-cfa4-6c48-8536-0279c867a896

+++ src/rti1516e/HLAfixedRecord.cpp
@@ -100,7 +100,7 @@ public:
     }
     uint8_t buffer[8];
     size_t offset = 0;
-    offset = encodeIntoBE32Compressed(buffer, 8, offset, length);
+    offset = encodeIntoBE32Compressed(buffer, 8, offset, static_cast<uint32_t>(length));
     offset = encodeIntoBE32Compressed(buffer, 8, offset, _version);
     return offset + length;
   }
